Add exit button in the shop
