+++ app/src/main/res/layout/activity_drawing_offline.xml
@@ -89,7 +89,7 @@
 
         <ImageView
             android:id="@+id/blackButton"
-            android:layout_width="0dp"
+            android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:layout_marginBottom="10dp"
             android:layout_marginTop="10dp"
